Simplify Alm and Healpix_Map interface.
Add routines ndarray2alm and ndarray2map to view an ndarray as Alm or Healpix_Map. It is not necessary to keep the intermediate arr since it doesn't own its buffer. Neither is keeping an intermediate ndarray, as long as the ndarray, which owns the buffer during the Healpix_Map or Alm lifetime is contiguous and of correct type.
Remove hack.h, not required.
Move Alm and Healpix_Map declaration in _common.pxd.