Use 'C' instead of 'Q' for equatorial coordinates
Aladin does not understand 'Q' for equatorial coordinates; it interprets
'Q' as galactic coordinates. Based on the HEALPix and Healpy source
code, it appears that the Java, C, IDL, and Fortran interfaces support
'C', whereas the C, Fortran, and Python interfaces support 'Q'.

This change affects just the FITS comment and the Python documentation
comments, so it doesn't actually restrict the user from setting
COORDSYS=Q, if necessary.