Use pykg-config if pkg-config is absent
If the program 'pkg-config' is not in the PATH, then pykg-config will be
downloaded (but not installed) and used as a replacement.

Note that 'pykg-config' will only be able to locate HEALPix or cfitsio
externally if their respective .pc files are on the PKG_CONFIG_PATH
because pykg-config will not have the same default search path as the
host platform's own native pkg-config would. This means that without
pkg-config most users will get the dependencies built from bundled
source.