Migrate from Distribute to Setuptools
Distribute was a fork of Setuptools, but the two have now merged into
Setuptools 0.7. This patch replaces the old Distribute bootstrap script
`distribute_setup.py` with the new Setuptools bootstrap script
`ez_setup.py`.

On systems with the new Setuptools, the build would fail with the
follwing error message:

    ValueError: A 0.7-series setuptools cannot be installed with
    distribute. Found one at /usr/lib/python2.7/site-packages

The new bootstrap script fixes this.