Fetch cfitsio and healpix_cxx as needed
...instead of including them in the release. This has the following
benefits:

 - Git submodules are no longer needed, so the Git repository is much
   more similar to a PyPI tarball.
 - Release tarballs are smaller.
 - Autotools is not required to build from Git.

Note that for the purpose of demonstration I have assumed that the
cfitsio and healpix_cxx tarballs are served from
`http://localhost:8000`. The primary cfitsio download URL is not
available today because of the U.S. government shutdown, and the
healpix_cxx download is not available because it has not yet been
decided where to host the Autotools-style tarballs.

To test, drop these tarballs into a single directory and fire up an ad
hoc web server in that directory with:

    $ python -m SimpleHTTPServer

and then run Healpy's `setup.py` as you normally would.