Bootstrap installation with a modern version of setuptools
Addresses #143.

Setuptools 0.6.10 is the oldest version with which we have tested
Healpy, and is also the least common denominator present on Scientific
Linux 6. If the user has setuptools >= 0.6.10, just take it. Otherwise,
let use_setuptools() download its default, more recent version.

While we are at it, update to the latest copy of ez_setup.py from the
Setuptools project.