Skip build-generated files in submodules when making source package
The directory `autom4te.cache` in the healpix_cxx sources is generated as
part of `autoreconf` and normally is not included when creating a tarball
(i.e., `make dist` doesn't copy it over). If the directory is present,
it causes some warnings when building the Debian package.

This patch causes the `autom4te.cache` directory to be skipped when
gathering files to be included in a source distribution
(`python setup.py sdist`).