Test several configuration simultaneously with travis
This rewrite of .travis.yml creates a matrix of eight build
configurations:

- Standard Ubuntu system Python versus Anaconda
- Install from source, or first build an sdist and install with pip
- Use system's external cfitsio library, or build bundled copy of
  cfitsio

Note that it sets `language: c` because the c language buildbot
configuration is a cleaner representation of a typical system. Also,
`language: python` does some very weird things to the system Python
path. Furthermore, this sets us up conveniently to use a Mac buildbot by
setting `language: objective-c` shortly.

Note that when using the system Python interpreter, this particular
configuration retrieves as many packages as possible with `apt-get`.