Adapt to some changes in behavior of pkg_resources.run_script
Previously, dependencies listed in `setup_requires` were available within
other scripts running inside the setup directory. Now that setup eggs are
downloaded to the `.egg` directory, it seems that we have to create a dummy
`Distribution` object to get to scripts.

Also, the `pykg_config.py` script had to be renamed to avoid a name collision
with the `pykg_config` package.

Fixes #216.