Convert /= operations to comply with Numpy 1.10.0 casting rules
According to the Numpy changelog
(http://docs.scipy.org/doc/numpy-dev/release.html):

> Default casting for inplace operations will change to ‘same_kind’
> in Numpy 1.10.0.

This is indeed the case, and Numpy inplace operations such as `/=`
now must have results that can be cast to the same size and type
as the left-hand side.

This patch is a straightforward transformation of the inplace operations
in `healpy.ud_grade`.