Fix readthedocs by disabling broken healpy_projaxes.rst (#374)
Readthedocs is hanging and/or erroring out because, while processing
`healpy_projaxes.rst`, it is trying to document not only the classes
in `healpy.projaxes`, but also the members of those classes that are
inherited from `matplotlib.axes.Axes`. This fails while attempting
to generate all of Matplotlib's example plots.

It is definitely not our intention to add all of the inherited methods,
but I can't get sphinx to respect the `:no-inherited-members:` option
(http://www.sphinx-doc.org/en/1.5.1/ext/autodoc.html). I even tried to
get the `matplotlib.axes.Axes` classes to work by enabling Matplotlib
sphinx extensions. I couldn't get it to work.

For now, disable `healpy_projaxes.rst` so that the Readthedocs
documentation builds. We can revert this when we figure out how
to fix the documentation for the `healpy.projaxes` submodule.

Fixes #336.