Add no_plot Boolean variable to gnomview
This commit adds the Boolean no_plot variable to the gnomview function in visufunc.py. Setting the variable to True will close the automatically generated output figure, which is useful is gnomview is run in a loop i.e. when stacking a large number of maps. Without this option, healpy will consume large amounts of RAM and can crash jupyter notebooks.