fix: remove nonzero on 0d warnings when plotting
This line caused warning:

DeprecationWarning: Calling nonzero on 0d arrays is deprecated, as it
behaves surprisingly. Use `atleast_1d(cond).nonzero()` if the old
behavior was inte
nded. If the context of this warning is of the form
`arr[nonzero(cond)]`, just use `arr[cond]`.

So instead of calling `where`, I just do mask==False,
it should also be possible to use np.logical_not