Only build wheels for 64-bit archs (skip i686)
The cibuildwheel CI job is timing out. Skip wheels for 32-bit
archs (i686). These archs are very rare nowadays and it is not
common for our upstream dependencies to provide wheels for them
anyway. Numpy, for example, does not distribute i686 wheels.

This should significantly reduce the build time because the CI
job will only have to build half as many wheels. The skipped
i686 wheels were the slowest anyway, because it had to build wheels
for many of the dependencies too (including Numpy).